implement delete product

+++ src/services/product.service.js
@@ -1,4 +1,5 @@
 import { PRODUCT_DESCRIPTION_PROMPT } from "../constants/prompt.js";
+import { ROLE_ADMIN } from "../constants/roles.js";
 import Product from "../models/Product.js";
 import promptAI from "../utils/ai.js";
 import uploadFile from "../utils/fileUploader.js";
@@ -64,7 +65,19 @@ const updateProduct = async (id, input, files) => {
   });
 };
 
-const deleteProduct = async (id) => {
+const deleteProduct = async (id, authUser) => {
+  const product = await getProductById(id);
+
+  if (
+    authUser._id !== product.createdBy.toString() &&
+    !authUser.roles.includes(ROLE_ADMIN)
+  ) {
+    throw {
+      status: 403,
+      message: "Access denied.",
+    };
+  }
+
   await Product.findByIdAndDelete(id);
 };
